requirements.txt
GitPython
Pillow
accelerate

basicsr
blendmodes
clean-fid
einops
fastapi>=0.90.1
gfpgan
gradio==3.41.2
inflection
jsonmerge
kornia
lark
numpy
omegaconf
open-clip-torch

piexif
psutil
pytorch_lightning
realesrgan
requests
resize-right

safetensors
scikit-image>=0.19
timm
tomesd
torch
torchdiffeq
torchsde
transformers==4.30.2